ComponentImportProjectFeature: Introduce InMemoryFileContent helper
